requirements.txt
setuptools
cassandra-driver 
pandas 
python-dotenv
pyyaml
scikit-learn
Jinja2
deepchecks
-e .